Add hash combine / avalanche functions
